Split BaseModel and GraphTower
